requirements.txt
python-dotenv~=1.1.0
py-cord
audioop-lts
openai~=1.82.1
torch
torchvision
torchaudio
